+++ RELEASE-NOTES.md
@@ -1,6 +1,9 @@
 ## RELEASE NOTES
-### Version 2.64.36-organisation-complex-field-type
+### Version 2.64.37-organisation-complex-field-type
 **EUI-2087** - Search organisation support full UK address fields
+
+### Version 2.64.36-hide-generic-callback-error-message
+**EUI-2344** - Hide generic callback error message when there is a custom configured message provided by a Service
 
 ### Version 2.64.35-organisation-complex-field-type
 **EUI-2086** - Search organisation